requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
Flask==1.0.2
geographiclib==1.50
geopy==1.21.0
google==2.0.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.2.0
nltk==3.4.5
oauthlib==3.1.0
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==5.4.2
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
tweepy==3.8.0
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
XlsxWriter==1.2.8
